requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.2               # via kombu
appnope==0.1.2            # via ipython
backcall==0.2.0           # via ipython
billiard==3.6.3.0         # via celery
brotlipy==0.7.0           # via -r requirements.in
cachetools==4.2.0         # via google-auth
celery[redis]==5.0.5      # via -r requirements.in
certifi==2020.12.5        # via requests, sentry-sdk
cffi==1.14.4              # via brotlipy
chardet==4.0.0            # via requests
click-didyoumean==0.0.3   # via celery
click-plugins==1.1.1      # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via celery, click-didyoumean, click-plugins, click-repl
coreapi==2.3.3            # via -r requirements.in, drf-yasg
coreschema==0.0.4         # via -r requirements.in, coreapi, drf-yasg
decorator==4.4.2          # via ipython
defusedxml==0.6.0         # via odfpy
diff-match-patch==20200713  # via django-import-export
dj-database-url==0.5.0    # via -r requirements.in
django-cors-headers==3.6.0  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django-import-export==2.4.0  # via -r requirements.in
django-js-reverse==0.9.1  # via -r requirements.in
django-log-request-id==1.6.0  # via -r requirements.in
django-model-utils==4.1.1  # via -r requirements.in
django-webpack-loader==0.7.0  # via -r requirements.in
django==2.2.17            # via -r requirements.in, django-cors-headers, django-filter, django-import-export, django-js-reverse, django-log-request-id, django-model-utils, djangorestframework, djangorestframework-simplejwt, drf-yasg
djangorestframework-simplejwt==4.6.0  # via -r requirements.in
djangorestframework==3.12.2  # via -r requirements.in, djangorestframework-simplejwt, drf-yasg
drf-yasg==1.20.0          # via -r requirements.in
et-xmlfile==1.0.1         # via openpyxl
facebook-sdk==3.1.0       # via -r requirements.in
future==0.18.2            # via python-twitter
google-api-core==1.24.1   # via -r requirements.in, google-api-python-client
google-api-python-client==1.12.8  # via -r requirements.in
google-auth-httplib2==0.0.4  # via -r requirements.in, google-api-python-client
google-auth==1.24.0       # via -r requirements.in, google-api-core, google-api-python-client, google-auth-httplib2
googleapis-common-protos==1.52.0  # via -r requirements.in, google-api-core
gunicorn==20.0.4          # via -r requirements.in
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
idna==2.10                # via requests
inflection==0.5.1         # via drf-yasg
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r requirements.in
itypes==1.2.0             # via coreapi
jdcal==1.4.1              # via openpyxl
jedi==0.18.0              # via ipython
jinja2==2.11.2            # via coreschema
kombu==5.0.2              # via celery
markuppy==1.14            # via tablib
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via -r requirements.in, requests-oauthlib
odfpy==1.4.1              # via tablib
openpyxl==3.0.5           # via tablib
packaging==20.8           # via drf-yasg
parso==0.8.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==3.0.8     # via click-repl, ipython
protobuf==3.14.0          # via google-api-core, googleapis-common-protos
psutil==5.8.0             # via -r requirements.in
psycopg2==2.8.6           # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
pygments==2.7.3           # via ipython
pyjwt==2.0.0              # via -r requirements.in, djangorestframework-simplejwt
pyparsing==2.4.7          # via packaging
python-decouple==3.3      # via -r requirements.in
python-twitter==3.5       # via -r requirements.in
pytz==2020.5              # via celery, django, google-api-core
pyyaml==5.3.1             # via tablib
redis==3.5.3              # via celery
requests-oauthlib==1.3.0  # via -r requirements.in, python-twitter
requests==2.25.1          # via -r requirements.in, coreapi, facebook-sdk, google-api-core, python-twitter, requests-oauthlib
rsa==4.6                  # via google-auth
ruamel.yaml.clib==0.2.2   # via ruamel.yaml
ruamel.yaml==0.16.12      # via drf-yasg
sentry-sdk==0.19.5        # via -r requirements.in
six==1.15.0               # via click-repl, google-api-core, google-api-python-client, google-auth, google-auth-httplib2, protobuf
sqlparse==0.4.1           # via django
tablib[html,ods,xls,xlsx,yaml]==3.0.0  # via django-import-export
traitlets==5.0.5          # via ipython
uritemplate==3.0.1        # via coreapi, drf-yasg, google-api-python-client
urllib3==1.26.2           # via requests, sentry-sdk
vine==5.0.0               # via amqp, celery
wcwidth==0.2.5            # via prompt-toolkit
whitenoise==5.2.0         # via -r requirements.in
xlrd==2.0.1               # via tablib
xlwt==1.3.0               # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
